[hqltables] Adds the Entity_Alias column to ad_column
In case that the the table has an alias in the HQL it is mandatory to set it on this field. This alias is used to add some mandatory filters on the Where clause replacing the @additional_filters@ string. Some examples of added filters are: client, organization and active flag.