Fixes issue 26231: Auxiliary Inputs are not computed in the Tree Grid View
Now when a record is selected in the tree grid view, the toolBar.updateButtonState() is called. This results in a FIC call in SETSESSION mode, that loads the auxiliary inputs and session attributes.

Also some changes have been done to ensure that if the tree grid view is shown, its selected record is used instead of the record selected in the standard grid.

--HG--
extra : rebase_source : 01ea2e88c240e5e74c2da9e8d895408527436e1a