Related to issue 26231: getParentRecord returns the proper record
The getParentRecord was returning the record selected in the standard grid view of the parent tab, even if it was showing the tree view grid. This resulted in that if the subtab had only one record, the tab display logic was not working properly.

--HG--
extra : transplant_source : %B7%23%29%99y%94%BD%2CR%7E-%C2%28%B7%9F%5DD%DF%AA%DC