Fixes issue 26262: Dates with negative offsets are formatted properly
The proper javascript format for dates and datetimes with timezone offset if the following: 2009-12-06T15:59:34+01:00. There if a method in charge of adding the semicolon in the datetime part in case it is missing.

Due to a bug in that method, the semicolon was not being added in dates with negative timezone offset. This resulted in a failure in the parsing of that date in the client side, being the date object actually modified when it was parsed.

In Openbravo there is a safety mechanism that detects if a record has changed based on its updated field. The updated field was being improperly modified in the client, so the safety mechanism was assuming that the record had actually been changed by another user, and was preventing the record from being saved.

--HG--
extra : rebase_source : b3aeb7c405b75f27946be32a89b0ba86e7e00ff1