Related to issue 26228: Date and datetime fields are shown properly in treeview
If a date or datetime field was shown in the tree grid view, the record was shown properly. But if then the record was opened in form view, the value of its date and datetime fields was being deleted.

--HG--
extra : transplant_source : %9FJ%80%0D%3F%0A%5C%F2%17fI%82d%84%D5%7F%D3%2A%C3%BF