Fixes issue 26286: Wrong behaviour if same column name exists in tab and subtab
This problem was reproducible if:
- There is a column in a subtab with the same column name as another column in its parent tab
- The column of the parent tab is flagged as stored in session
- The column of the subtab is not shown in the grid
- The value of the record in the subtab has a null value in that column

If all these conditions are met, then when the record in the subtab is opened in the form view, its null value will be replaced with the value of the column of its parent tab.

This bug happens because the FIC does the following:
- It invokes setSessionVariablesInParent. This method stores in the RequestContext the session variables of the parent tab, among others. At this point the stored in session column of the parent tab is stored in the RequestContext
- It invokes setValuesInRequest. This method stores in the RequestContext the values of the row being opened in the form, but only if does it for the values that are not null. In this case, the value is null so it is not stored in the RequestContext, so then when the RequestContext is accessed to fetch the column values, the value stored in the setSessionVariablesInParent method will be returned.

The fix only affects to the FIC calls in EDIT mode. In these calls the actual value of the record should take precedence over the values stored in session.