[hqltables] Implement the HQL Query Transformer
An HQL Query Transformer can be used to transform the hql query of a hql table. HQL Injectors can only inject new code at predefined points, but the HQL Query Transformers can modify the query at any point.

For instance, this snipped of a HQL Query Transformer adds a column to the SELECT clause of the query:

  public String transformHqlQuery(String hqlQuery, Map<String, String> requestParameters,
      Map<String, Object> queryNamedParameters) {
    String transformedHql = hqlQuery.replace("from ", ", e.symbol as em_hqlext_symbol from ");
    return transformedHql;
  }