Fixes issue 26420: Cannot modify an existing AD Tree Table Tree Category record
The problem was that there was an event handler on the ADTableTree table that checked that at most one ADTree tree is defined for a particular table at all times.

The logic that handled this checked on record creation and edition that no ADTree trees were already defined for that table. But there was a bug: if an ADTree tree was modified, then that same tree was detected as an existing ADTree tree, and the event handler returned with an exception.

To fix this now the event handler takes into account the id of the ADTableTree being modified when looking for exisint ADTree trees.

--HG--
extra : rebase_source : b5e999be0cb31e062e56bc8a8320660f7aa2d8d2