[combos refactor] force fetch in few cases
We invalidate the cache and force to fetch data from datasource in the following cases,
for new records, in case the column
has been changed from FIC either through validation rule or from related columns,
but does not contain entries.
Also when editing a record, if no fetch was done before, server fetch is done.