[pw3] Hides properly the filter editor bar if there are no filterable fields
The filter editor bar will be shown if the grid has a filter clause, even if it does not have any filterable fields. This has to be done in order to show the funnel icon to clear the filters, otherwise it would not be possible to clear the filter clause.

Also, the filterEditorProperties is cloned to avoid sharing the same object with all the grids.