Fixes issue 26636: Error in int-checks-pgsql
Before a refactor, the showTreeGrid attribute was declared, initialized and used in a function. After this refactor, the attribute is declared and initialized in a function, and used in another, without being declared in the latter function.

To fix this, now the first functions passes the showTreeGrid attribute as an argument to the second function.