Fixes issue 26820: Additional properties should not be used as primary keys.
When setting the primary keys of a datasource, its additional properties should not be taken into account. Only those that belong to the main entity should be considered.

For instance, consider the following case. The SalesInvoiceLine entity has a product property, which uses a selector reference. The field used as the Display Field of the selector will be added to the additional properties of SalesInvoiceLine. If the product.id property were used as Display Field, then it would be added to the additional properties, but should not be considered a primary key for SalesInvoiceLine even though that property is a primary key in its referenced entity.