Fixes issue 23889: Multiple requests to selector datasource in manual popups
Pick lists and grids use a property called drawAllMaxCells, that is used to determine how many cells should be fetched from the datasource, its default value is 250. If after making a datasource call smartclient finds out that the number of retrieved cells is lower than drawAllMaxCells, it will do another datasource call, and repeat this process until the number of cells fetched are higher than drawAllMaxCells or until all cells have been fetched.

This is not the intented behaviour, we want the rows to be fetched when the user scrolls down and is out of cached rows. This is achieved by setting drawAllMaxCells to 0.