26696: Table references work with display column other than identifiers
When table references were displayed, always the identifier columns were displayed.
The actual value was displayed in cases of composite identifiers which in turn was another identifier,
eg., Financial account in Customer. Now it uses the display column in all cases.
The criteriaDisplayField parameter is already being used for displaying the appropriate value. Set it's value correctly for table references
Added additional logic in DefaultJsonDataService to include the display field column to the formed BOB object.
This fix is not risky as it affects only the displaying of FK filter values from the fetched result.