Fixes Issue 26838: Unpaged request in Financial Account Transaction
Perform form refresh after completing button action only if current form is valid.
In the issue, we are deleting the record using the button, so the form present after the delete action is invalid
resulting in fetch with incorrect critieria to the server.
The problems happens only in form view, as refresh grid is done with proper criteria.