Fixed issue 26397: Cash VAT for orders and invoices based also on location
Cash VAT regime is only allowed for transactions inside the same country.
In case the Organization's and Business Partner's countries are different, the document's Cash VAT flag (for orders and invoices) is forced to be unchecked.
This is controlled through several callouts at Invoice's/Order's header on the Organization and Business Partner Location columns.