[fkfilters] Client filtering works even if picker dropdown is not loaded
The picker dropdown was being used to obtain the record id based on its identifier and viceversa. But sometimes this action needs to be done before the picker has been loaded. For this, a new cache has been added, called filterAuxCache. If the object being looked for is not in the picker drop down, then it will be searched in this cache.