[fkfilters] Adds support to Saved Views
When a view is saved while any of its filters are currently using the 'id' filter type, the id and identifier of the selected options must be stored in the cache. This way the cache can be loaded when the view is applied. This is needed because at that point the filter dropdown has not been loaded yet, so no other mapping between the id and the identifier is available.