Fixes issue 26696: Filters of Table references work properly.
When table references were displayed, always the identifier columns were displayed.
The actual value was displayed in cases of composite identifiers which in turn was another identifier,
eg., Financial account in Customer. Now it uses the display column in all cases.
The identifier is set with appropritate value in filters also.