[fkfilters] Implements logic for allowFkFilterByIdentifier grid configuration
If the allowFkFilterByIdentifier grid configuration is set to false, then the only filter type allowed will be 'id'. If the user manually updates a fk filter with this configuration, when the focus leaves the field no filters will be done, and the criteria for that field will be reverted to its previous one.

Even if allowFkFilterByIdentifier is set to false, the 'identifiler' filter type will be used when populating the filter drop down.