[fkfilters] Adds KernelUtils.hasNullableIdentifierProperties utility method
Given an entity, this method returns true if at least one of the properties that are part of its identifier is not mandatory, false otherwise.

This method will be used to know if a 'inner join' can be used instead of a 'left join' when filtering a table using a foreign key column (see issue https://issues.openbravo.com/view.php?id=26279).