[fkfilters] Users 'inner join' instead of 'left join' when possible
When creating a query as a result of filtering the grid using a foreign key column, it is better to use a 'inner join' instead of a 'left join' to join the main entity with the entity targeted by the filtered column.

This cannot be done always. These two conditions must be met in order to be able to user 'inner join':
- All the columns that are part of the identifier of the target entity must be not nullable (mandatories)
- The joined entity can only be used in the where clause, due to filtering the grid. For instance, if the Sales Order grid is filtered by the Business Partner column, and also being sorted by it, a 'left join' will be used. The gridFilterExclusiveJoinMap attribute is used in the AdvancedQueryBuilder class to know if the joined entity is being used only in the where clause or not.