Fixes Issue 27061: Sales Representative field is not initialized properly
In the combos project, the FIC always returned one value when opening/editing a record.
To achieve this, whenever columnValue was present it was used. But there are cases where columnValue is the session value.
So setting the first element only if it matches with the columnValue now. Or else removing the value as it is not needed.
When the dropdown is clicked, data is fetched from ComboTableDatasourceService, so it is not affected with this change.