Related with issue 25211: Adds Grid Config to QA Testing sampledata (updated)
A Grid Configuration setting has been added to theQA Testing sample data in order to test the FKPERF040_FilterForeignKeyColumnWithIdentifierFilteringDisabled test. The Grid Configuration is used in the FKPERF040_
FilterForeignKeyColumnWithIdentifierFilteringDisabled test.

The Grid Configuration added does not allow to filter using the identifier in the Visible at Client field of the header tab of the Preference window.

Now the grid configuration is set to use the System client.