Fixes issue 27017: Callouts of combos on manual forms does not work properly
The original problem was that the response of the callout was not being applied to the proper iFrame. To solve this, the activeFrame attribute was created to know which frame is active at any given time, in order to apply the callout response in that frame.

The problem of this issue is that the activeFrame was not being stored properly when opening a new manual window.