Fixes issue 27303: direct link filtered by FK column doesn't work
The problem was that the identifier of the foreign key being filtered now is not stored in the criteria itself, because it is filtered using the id. To fix this, the chache with
the map of fk-id, fk-identifiers of the filtered foreign keys is stored in the url, and then applied when the grid filter editor is available.