Fixes issue 27297: Bad behavior when a record is created with lazy filtering
The problem was that the grid data, a ResultSet, was not being created until the first fetch call was done. If a record is created before the first fetch call, which is possible only using lazy filtering, then the record will not be shown in the grid until a fetch call is done, although it will be stored properly in the database. This also results in a wrong message being shown in the active subtab after saving the record.

To fix this, and empty ResultSet is created when a new record is created if the grid does not have a ResultSet yet.

--HG--
extra : rebase_source : 9db7663b72ef0dc1010d30692e7206b42c799861