Fixes issue 27246: Error when setting count function in analytics window
The problem was that a select clause like this was being used in a HQL query:
* select count(e.salesOrder.documentNo + join_1.orderDate + join_1.grandTotalAmount) from ...

Property concatenation is not supported in a count function, it does not even make sense. To fix this, the addSelectFunctionPart method of AdvancedQueryBuilder has been updated so that '*' is always used inside the count function.

--HG--
extra : rebase_source : 627fee0e34ae1513ddb2bb5ba737557da2012302