Fixes issue 27297: Bad behavior when a record is created with lazy filtering
There was one case that was not being considered: Clicking the 'Apply filters' icon just after creating a new record without any criteria applied. It was not working because it was expecting the grid data not to be a ResultSet, but now it is a ResultSet.

To fix this, a new flag has been created to specify when the grid has yet to load its data.

--HG--
extra : rebase_source : e5393d14522cc5215b1fcf4436b27932bca4aad3