Fixes issue 27405: Problems filtering P&E grids in slow environments
The problem was that when the response of the first datasource request was received, the second one was already in progress, so the grid local data only contained 'loading' records. At that point the grid local data was not accessible. To fix this, now it takes into accout if the grid local data can be accessed.

--HG--
extra : transplant_source : %12WN%BF%96%3A%9A%8A%9A%84%DA%1C%F8M%CF%A1%F28M_