Fixes issue Display Logic 27122: in fields does not work with property fields
The problem was that if a property field is used in a display logic, it was going to be included in the sessionAttributes in the call in EDIT mode done to the FIC, and its value was going to be reset to an empty string. To fix this, a field is not added to the list of session attributes if it is a property.

This change only takes effect when using a property field in a displayn logic, so it should not have any negative effect.

--HG--
extra : rebase_source : b304b79fa0773d78e2dd7ec70c624e76560b6b81