[reconciliation refactor] Include context of outer parameter window
When a parameter window is opened from another parameter window, the parameters of the outer parameter window were not being sent as request params in the initialization of the inner parameter window. Now all the non-grid parameters are sent. The grid parameters are not sent because they can contain so much data that they can break the request.