Fixes issue 27066: Dojo autocompleting changes the date separator
In the autocompleting process, dojo needs to use the '-' date separator. The problem was that after doing the autocompletion it was not respecting the original date separator, so if it used '/', this date 01/02/03 would to autocompleted to 01-02-2003.

To fix this the date separator is saved before purging the date format.

--HG--
extra : rebase_source : 20aafc7c3d495eefd131651ab548026c377a703f