Fixes issue 26457: Date columns not properly displayed in query/list widgets
The problem was that all date columns were being stored internally as DateTime column, so they were being treated as such and were being converted to the local timezone. As these dates at a time of 00:00:00, if the client whose timezone < the timezone of the server, the datetime shown was the previous one.

To fix this, Date columns are not sent as proper dates, instead of as DateTimes. This way the date columns will show the same value in all clients, regardless of their time zones.

--HG--
extra : rebase_source : 5b8b9522a87a9b892e03beab463726477cf52084