Fixes issue 27503:  Form view is not working fine when using views
The problem was that the OB.Personalization.updatePersonalizationDataFromFields functions, used in the process of storing the view, was not perfectly adapted to store forms that have not been initialized yet. The fields have not been created yet, so instead of a formitem only the main properties are available. This means that:
- isc.isA.SectionItem does not work, because it requieres a Class object. To fix this, the field type (already available) is also checked.
- Some attributes of the fields (rowSpan, colSpan and newRow) have not been initialized yet and unless its value is other than the default, they are undefined. To fix this they are given its default values (1, 1, false) when the view is stored.

--HG--
extra : rebase_source : 6cf7a4f2074b083041cf952760de137995437123