Fixes issue 27322: Param windows opened from the menu fill in _org parameter
The problem was that the organization used to set the _org parameter was the one of the record selected in the view from where the process window has been opened. If the window is opened from the menu then the _org parameter was not being set, and some validations were not working.

Now, if the parameter window has an ad_org_id parameter and is not empty, its value will be used to set _org. Otherwise the current user organization will be used.

--HG--
extra : rebase_source : da28a9b8056e4d26b9201178754a34be7003ea57