Fixed issue 27544: Visibility problems in some dropdown selectors.
Real problem is that is not possible to use the dropdown of a
selector when the line where the selector is located is on the
bottom of the page and the selector has more than one field to show.

To avoid this problem, it has been modified setPickListWidth() in
order to prevent a pickListWidth longer than the width of the grid.

In the other cases (the selectors have one field to show),
setPickListWidth() works correctly.