Fixes issue 27633: Tree view does not show Loading message when it has no data
The problem was that the noDataEmptyMessage attribute, that contains the message to be shown when the grid does not have any data, did not have the proper value. This attribute is reused to also show the Loading message while the datasource request is being processed, but was not being reset to its proper value (OBUIAPP_NoDataInGrid message) when the data arrived to the tree.

Now in the OBTreeViewGrid.dataArrived function the message is properly reset before invoking the resetEmptyMessage function, that will actually display the message if the datasource did not return any records.