Fixes issue 27630: The proper date format is used in Pick and Execute windows
When a datetime field was shown in a P&E window, the date format was used, instead of the datetime format. This happened because of this bug in the DataToJsonConverter.convertPrimitiveValue():

    ...
    if (value instanceof Date) {
      return xmlDateFormat.format(value);
    }
    if (value instanceof Timestamp) {
      return xmlDateTimeFormat.format(value);
    }
    ...

Timestamp is a subclass of Date, so all values that are instance of Timestamp are also instance of Date. This means that if the value attribute contains a datetime, the execution flow will enter in the first if statement, and the date format will be used.

This bug is fixed by checking first by checking first if the value is an instance of Timestamp before checking if it is an instance of Date.

This bug was not reproducible ins standard grids because this function is only called when using ReadOnlyDataSources.