Fixes issue 27621: Button works after having connectivity problems
The problem was that there was in place a mechanism to prevent accidentally opening a process twice, that was not working properly if the button was clicked while having network conectivity problems. In that case a flag (button.opening) was being set to true when the button was clicked, but was never set to false, so the next time the button was clicked the mechanism thought that it was still being opened and canceled all subsequent openings.

To fix this, a new parameter, errorCallback, has been added to the OB.RemoteCallManager.call function. This function, if defined, will be executed when the call returns with a status error.