Fixes issue 27636: Absolute Date Time field works properly with offset change
The problem was that if the date selected in an absolute datetime field was very close to a timezone offset change, the date was automatically changed. The original entered date was modified in the OB.Utilities.Date.addTimezoneOffset function. Now, if inside this function the date timezone offset is modified, the proper correction is applied.

--HG--
extra : rebase_source : 28e735ab6b0e5fe2abe3ef519a82e7cb5b08fef6