Fixes issue 27053: Numbers are properly formatted with , as decimal separator
Decimal numbers were not being properly formatted if a comma was used as decimal separator. The problem was caused by using the OBPlainToOBMasked instead of JSToOBMasked to convert a javascript number to a formatted string. OBPlainToOBMasked does not expect a javascript number as an argument, but a non-formatted string representation of a number. The issue has been fixed by replacing OBPlainToOBMasked with JSToOBMasked.

The issue that caused this regression [1] works the same as it worked when it was fixed. It works properly using the '.' as decimal separator, but not when using ','

[1] https://issues.openbravo.com/view.php?id=25517

--HG--
extra : transplant_source : %F5%CA%AD%15%82a%16%8BCG%C1%F9%2B%BA%FA0%1E3%04%E7