Fixes issue 27706: fk filters work having several records with same identifier
The problem was that the new way if filtering foreign keys by using their ids instead of their identifiers did not support having several records with the same identifier. Note that this is a corner case, as it should not happen that two records in the same table have the same identifier.

To fix this, the getRecordIdFromIdentifier function has been reworked to getRecordIdsFromIdentifier, because a given identifier can hit several records, each one with its own id. Now, if for instance the user clicks on the identifier x, and there are several records with the x identifier, all those records will be checked in the picklist.