Fixed bug 26826: Allow Period Control invalid for organization type
The Allow Period Control flag is only shown in form view when the organization type is either a legal entity with accounting or a business unit. However the user is able to overwrite the Allow Period Control flag using the grid view, where the display logic doesn't work. This situation can lead the user to set this flag for incorrect organization types, like "Organization" and "Generic".

To fix this issue we have included a validation into the AD_ORG_TRG and we have created a module script that set the Allow Period Control to N when the associated organization type is neither a legal entity with accounting nor a business unit.