Fixes bug 27709: Inactive Window/Tab/Field Access records are not considered
The problem was that even if a WIndow/Tab/Field access record was inactive, it was being taken into account by the WindowSettionsActionHandler class. Now this no longer happen.

This fix comes with a behavioural change in the way the Is Active field of these tabs is handled. Before this fix the flag was ignored, now is taken into account.