Fixes issue 27710: Fixes problem defining Field Access for process buttons
The problem was that if several process buttons were defined as not editable in the Field Access tab, then in the subtabs of the tab that own the issue only one of them was being set as read only. This happened because when the button being set as read only was looked for in the subtabs, the property of the button in the main tab was not being compared with the property of the button in the subtab, so the first one disabled was picked.