Fixed bug 26893: Wrong warehouse when creating a sales order from an expense.
An auxiliary input was added to associate the created expense sheet warehouse with the selected organization's warehouse.
Now when Create Sales Orders from Expenses process is launched, sales order is associated with the correct expense's warehouse.
Also the process throws an error message if the expense's organization does not have a default warehouse defined.