Fixes issue 27730: Pressing enter in the last editable grid line creates record
The fix [1] of this issue [2] added a check to prevent creating a new line when, in an editable grid, a value was selected from a picklist by pressing the Enter key. The fix consisted in overwriting the ha
ndleKeyPress function of TextItem. The problem was that because of the way the fix was implemented, the new function was invoked but the original implementation of TextItem.handleKeyPress was never execute
d. This missing code was causing a strange behaviour in how the EventHandler set the items that are object of the events, and as a result of this the current issue was being reproducible.

To fix this, the issue [2] has been implemented differently. The root cause of that issue is that the keyDown event was done in a formitem (the selector) and the keyPress on another item (in this case the
description column). When this was detected, the call the handleKeyPress was intercepted to prevent adding a new line. Now, the interception is done in the cellEditEnd. This has two benefits:
- handleKeyPress is not overwritten, so the problems caused by the missing TextItem.handleKeyPress invocation disappear
- the interception is done in a more specific point. handleKeyPress is called constantly, but the cellEditEnd is only called when the focus leaves a grid cell.

[1] https://code.openbravo.com/erp/devel/pi/rev/b4ba48792e4850c804c42001d8f818f1e0575f81
[2] https://issues.openbravo.com/view.php?id=26817

--HG--
extra : transplant_source : %FF%95%C8%C7h%E7%99%DBO%7D%8E%A1%D4%3E%DD%1E%DFy%3B%EB