Fixes issue 27751: discounts is enforced to be dependent of mobile.core
a. If the discounts module is dependent of mobile.core, the ERP is forced to install mobile.core, which is unwanted.
b. If the discounts module is not dependent of mobile.core, the javascript files are inserted before the mobile.core javascript files. This behavior prevent discounts to be computed

A hack has been inserted until better decoupling of mobile from the ERP is made