Fixed bug 26562: Error message Unprocessing TaxPayment related to posted Payment
Improved the error message shown when unprocessing a Tax Payment related to a posted Payment.
Besides, the Payment field is shown when the Tax Payment is related to a payment to make it easier to identify the payment.
Finally the "Generate Payment" field is always shown because it had no sense to hide it when the record was processed.