Fixed bug 27753: Tax Payment Unprocess when Generate Payment = Y (ORACLE only)
The extension point when unprocessing a Tax Payment didn't work in Oracle, reason: a SQL query was expecting to receive always a GL Item, but when unprocessing a record, the GL Item is not sent to the extension point. This created an Oracle exception that wasn't properly captured.

The solution is to move the SQL query that requires the GL Item inside the processing flow only.