Fixed bug 27775: Performance improvement in 2 module scripts
The TruncInvoiceDate and Issue26826_Org_AllowPeriodControl module scripts had the same performance problem: they ran a full table scan each time the module script was executed regardless it updated any record or not.

To fix it, the module script now verifies whether a preference exists or not. If it doesn't exist, the update is executed and the preference is set, so the update process won't be executed in the future.