Fixes issue 27769: Record id not shown in the filter when a new field is added
When a foreign key field is being filtered using its id (i.e. by selecting a record from its drop down), its filterType property is set to 'id'. This property is then used in several places like in the getCriterion and getCriterion. The problem was that when adding/removing a field from the grid, the filter fields are rebuilt from scratch and the filterType property was being reset to its default value ('identifier'). To fix this, the list of foreign key fields being filtered by their id is stored before resetting the filter fields, and consulted when initializing the foreign key fields.

Also the foreign key cache needs to be persisted in order to be able to obtain the record's identifier given its id (in the criteria only the id is stored).

The getFKFilterAuxiliaryCache function has been moved from OBViewGrid to OBGrid so that it can also be used in Pick&Execute grids. This has no risk, as it did not use any code specific from OBViewGrid.

--HG--
extra : rebase_source : f89ee1fb4848881ef5cc707af607a57f74ab3152