Fixes issue 27769: Record id not shown when opening link from URL
In the OBGrid.setFields function, the originally named fkCache was meant to store the information needed to restore the state of the foreign key columns after resetting the fields. The scope of that auxiliary cache was setField function, that is why it is deleted from the object before leaving the function.

The problem was that 'fkCache' was already being used to name another property. This property stores the id-identifier pairings of the criterias saved in direct links. As that property was being removed when the grid fields were set, if a tab was opened using a direct link the criterias on the foreign key columns showed their ids.

This problem has been fixed by renaming the temporal property used in the setFields function.

--HG--
extra : transplant_source : q%98%F6%B0P%D3%8D.%02%DB%06/%9F%19.%804K%F2%9E