Fixes issue 27768: Criteria built properly for rows with non unique identifiers
The problem was that when building the list of ids to be used in an id-filter, the list could contain duplicate values.

Consider the following case: There are two business partner with the same identifier. When one of them is selected in the filter drop down, both of them will actually be filtered, due to having the same identifier. When the criteria was being built for this field, each checkbox checked produced 2 ids, because there were two records with different ids but with the same identifier. If both records are selected, then each of them would add the two ids, resulting in four ids.

This has been fixed by making sure there are no duplicate values in the list of ids.

--HG--
extra : rebase_source : 8eec4e76a95ce8b572f0a59366dccc34167032da