Related with issue 27788: Adds new field to obserds_datasource table
The new field is a flag that controls how the foreign keys are filtered in the tables that uses a manual datasource to fetch its data. If the flag is checked, then if a foreign key field of that table is filtered by selecting a record from the filter drop down, the foreign key will be filtered using its id. That is, the criteria will be like this:

  {fieldName: foreign-key-name,
  operator: 'equals',
  value: record-id}

If this flag is unchecked, then the criteria used will be like this:

  {fieldName: foreign-key-name$_identifier
  operator: 'iContains',
  value: 'identifier'}