Fixed bug 27872: NPE when reactivating payment for BPartner without currency
When reactivating the payment, the FIN_PaymentProcess was getting the business partner currency's ID where, in some strange circunstances, it may not be defined yet.
Moreover, getting the BP currency ID at this point is not needed at all, so these lines have been removed.

Finally, in another part of the code, we have done a trivial refactor to use Currency objects instead of Strings to have a better control over NPE.