Fixes issue 27912: Time&DateTime fields are properly handled in P&E Windows
The values of Time and DateTime fields were not being properly handled, as they where not converted from the server timezone offset to UTC before being sent to the client. This happened
because the DataToJsonConverter.convertToJsonObjects method did not perform the conversion.

Now it does, when possible. For the conversion to be done properly, the Property associated with the column value must be passed to the convertPrimitiveValue method. This will be done a
s long as the tableId parameter is sent in the datasource request, and as long as the property name can be inferred from the column name.

The QueryListDataSource extends ReadOnlyDataSourceService, but the tableId is not passed, so this changeset does not fix this case, that will be address in a separate issue.

--HG--
extra : transplant_source : %9055%114%40%7D%BC%BE%0B%02%82n%EE%BC%91%0E%DFt%B5