Fixed issue 0027921: Layaways are not being invoiced since Q2 cash up oflline change
When a layaway is created we assign a cashupid, but if we payfull or partially again this layaway in another cashup, this transactions will not be reconciliated and invoice will not be created. We need to update the layaway with the new cashupid.

--HG--
extra : transplant_source : %0B%E6%BE%C9%C7%C8L%40%02%A3J%82%E3%11%94%23%E3%B23%C3