Fixes issue 27918: User Defined Accounting Report Setup shows tree view
The problem was that the tree configuration previous to the new tree implementation was wrong. The Table table had a field called Type Area which was originally used to specify the Tree type of the table. The FinancialMgmtAccountingRptElement table had this field empty, and instead the FinancialMgmtAccountingReport had this field set to Accounting Report, which is the value that should have been assigned to FinancialMgmtAccountingRptElement.

The configuration of the new implementation was done based on the Tree Type field, so it was wrong for the FinancialMgmtAccountingRptElement table. Now it has been properly done (it has been flagged as Is Tree, a record has been added to the Table Tree Category subtab and a value has been selected in the Table Tree Category field of the Tab tab.

Also, a module script has been modified to fix the data of ad_tree. The ad_tree table has a column called ad_table_id, which points to the table that the tree belongs to. The accounting report tree was mistakenly pointing to FinancialMgmtAccountingReport, and the updated modulescript fixes it so that now it points to FinancialMgmtAccountingRptElement.

--HG--
extra : transplant_source : %20%3By%7EA%89SOK%1D%FC%A9bX%8E%3D%60%13%ABb