Fixes issue 27918, fixes issue 27930. Prevents breaking the int-inc tests
The problem was that due to this design defect [1], the oracle and postgres int-inc tests were failing. This is because a value 'AR' was being set to the treetype of a table and then being set to null in other table whose treetype was also 'AR'. Treetype has a unique constraint that was enabled when the application data was updated, so the constraint failed.

To prevent this error, the treetype of the second table is going to be set to null in a modulescript, that is executed before updating the application data.

[1] https://issues.openbravo.com/view.php?id=12577

--HG--
extra : transplant_source : %8A%23%D6n/k%974%C1%40v%E8%60%94%CCRz%C1%04L