Fixed bug 27461: Sales invoice Dimensional Report improvements for Credit Memo
In case of credit memo documents, the following columns in the report should reflect a negative number:
 Quantity
 Amount
 Cost
 Profit
 Margin

Before this fix, only the Amount column reflected a negative number. Besides the Profit and Margin were wrongly calculated in credit memo scenario, because they are based on the Cost which had a wrong possitive sign for credit memos.
After this fix all these columns show a negative number in case of credit memo. Thus the Profit and Margin are properly calculated based on the right sign of the Cost column.

Note that the Margin formula has been also updated to reflect a negative sign in case the Profit is also negative:
Profit = Amount - Cost
Margin = 100 * ((Profit) / Amount) * sign(Profit)

That means the Margin will now be shown as a negative percentage for credit memo and negative invoices.