Fixes issue 27918: Removes tree structure from table that is not a tree
In the initial changeset for fixing 27918, the tree type of the FinancialMgmtAccountingReport table was set to null because that table is not really a tree. The problem was that the cleaning up of the table was not complete, the Is Tree flag should be set to false, and its Tree Table Category subtab should be empty. This has been done now.

The Initial Client Setup was failing because of the incomplete cleanup. The Tree Table Category of FinancialMgmtAccountingReport had a record called Accounting Report, which is the same name as the tree of the FinancialMgmtAccountingRptElement table.

--HG--
extra : transplant_source : _%9F%3D%AEn6a%A7%9E%E3%93%3EoEI%94%D3%2BC%A7