Fixes Issue 27266:United States is always selected in Country field (Business Partner window, Bank account tab).
Default value shown on country field in Bank account tab (Business Partner
window) is always United States and it should not be in that way.
Now, default value will be the country defined in Location/address tab.
If there are no Locations defined, it will take the country which has isDefault
column set as 'Y'. Finally, if there are several Locations defined for that
Business Partner, the system will take any of them and if there is one record
whose country is set as Default, this one will be shown.